@@ -24,8 +24,12 @@
   ],
   "license": "GPL-2.0-or-later",
   "require": {
-    "php": ">= 5.3.3, <= 7.1.99",
-    "typo3/cms-core": ">= 4.5.0, <= 7.99.99"
+    "php": "^5.6 || ^7.0 || ^7.1 || ^7.2",
+    "typo3/cms-core": "^7.6 || ^8.7 || ^9.1",
+    "ext-openssl": "*"
+  },
+  "suggest": {
+    "typo3-ter/openid": "^7.6 || ^8.0"
   },
   "autoload": {
     "psr-4": {
