@@ -10,7 +10,7 @@
  * writing. "version" and "dependencies" must not be touched!
  ***************************************************************/
 
-$EM_CONF[$_EXTKEY] = array(
+$EM_CONF[$_EXTKEY] = [
     'title' => 'Central Account Management',
     'description' => 'This extension allows TYPO3 administrator accounts to be managed centrally and automatically synchronized with remote websites using a secure link. No need for complex LDAP / ActiveDirectory infrastructure.',
     'category' => 'services',
@@ -18,9 +18,6 @@ $EM_CONF[$_EXTKEY] = array(
     'author_company' => 'Causal Sàrl',
     'author_email' => 'xavier@causal.ch',
     'shy' => '',
-    'dependencies' => 'openid,scheduler',
-    'conflicts' => '',
-    'suggests' => '',
     'priority' => '',
     'module' => '',
     'state' => 'stable',
@@ -30,16 +27,16 @@ $EM_CONF[$_EXTKEY] = array(
     'modify_tables' => '',
     'clearCacheOnLoad' => 0,
     'lockType' => '',
-    'version' => '1.7.1',
-    'constraints' => array(
-        'depends' => array(
-            'php' => '5.3.3-7.1.99',
-            'typo3' => '4.5.0-7.99.99',
-            'openid' => '',
+    'version' => '2.0.0-dev',
+    'constraints' => [
+        'depends' => [
+            'php' => '5.6.0-7.2.99',
+            'typo3' => '7.6.0-9.1.99',
             'scheduler' => '',
-        ),
-        'conflicts' => array(),
-        'suggests' => array(),
-    ),
-    '_md5_values_when_last_written' => '',
-);
+        ],
+        'conflicts' => [],
+        'suggests' => [
+            'openid' => '',
+        ],
+    ],
+];
